requirements.txt
pydantic>=2.7
requests>=2.32
pyyaml>=6.0
tqdm>=4.66
